Fix auto resizing

+++ src/main/java/org/poianitibaldizhou/sagrada/graphics/utils/SceneManager.java
@@ -71,11 +71,11 @@ public class SceneManager {
         return (Stage) stackPane.getScene().getWindow();
     }
 
-    public int getSceneWidth() {
+    public double getSceneWidth() {
         return windowSize.getWidth();
     }
 
-    public int getSceneHeight() {
+    public double getSceneHeight() {
         return windowSize.getHeight();
     }
